services/predictor/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
aiofiles==23.1.0
celery==5.2.7
Jinja2==3.1.2
pytest==7.2.2
redis==4.5.4
httpx==0.23.3
